requirements.txt
Django==4.2.15
djangorestframework==3.15.2
django-debug-toolbar==4.4.6
drf-yasg==1.21.7
